requirements.txt
numpy==1.18.5
scipy==1.4.1
scikit-image==0.17.2
matplotlib==3.3.4
Pillow==8.3.1
ipython==7.16.1